chore: remove legacy single-file assets
