+++ .prettierrc
@@ -1,0 +1,4 @@
+{
+  "singleQuote": true,
+  "tabWidth": 2
+}
