+++ README.md
@@ -1,1 +1,21 @@
 # KP-generator
+
+## Development
+
+1. Install dependencies
+   ```bash
+   npm install
+   ```
+2. Start development server
+   ```bash
+   npm run dev
+   ```
+3. Build for production
+   ```bash
+   npm run build
+   ```
+4. Lint and format source
+   ```bash
+   npm run lint
+   npm run format
+   ```
